Move network related DI components into their own package
