+++ app/src/main/java/org/simple/clinic/di/AppModule.kt
@@ -11,6 +11,7 @@ import dagger.Provides
 import org.simple.clinic.appconfig.AppConfigModule
 import org.simple.clinic.appupdate.AppUpdateModule
 import org.simple.clinic.crash.CrashReporterModule
+import org.simple.clinic.di.network.NetworkModule
 import org.simple.clinic.facility.change.FacilityChangeModule
 import org.simple.clinic.illustration.HomescreenIllustrationModule
 import org.simple.clinic.login.LoginModule
