+++ app/src/main/java/org/simple/clinic/di/network/RetrofitModule.kt
@@ -1,10 +1,12 @@
-package org.simple.clinic.di
+package org.simple.clinic.di.network
 
 import com.squareup.moshi.Moshi
 import dagger.Module
 import dagger.Provides
 import okhttp3.OkHttpClient
 import org.simple.clinic.BuildConfig
+import org.simple.clinic.di.AppScope
+import org.simple.clinic.di.CountryRetrofitModule
 import retrofit2.Retrofit
 import retrofit2.adapter.rxjava2.RxJava2CallAdapterFactory
 import retrofit2.converter.moshi.MoshiConverterFactory
