+++ app/src/test/java/org/simple/clinic/sync/LastSyncedStateTest.kt
@@ -2,7 +2,7 @@ package org.simple.clinic.sync
 
 import com.google.common.truth.Truth.assertThat
 import org.junit.Test
-import org.simple.clinic.di.NetworkModule
+import org.simple.clinic.di.network.NetworkModule
 import org.threeten.bp.Instant
 
 class LastSyncedStateTest {
